LayoutTests/http/tests/paymentrequest/payment-address-attributes-and-toJSON-method.https-expected.txt

CONSOLE MESSAGE: WebKit currently uses the first shipping option even if other shipping options are marked as selected.
CONSOLE MESSAGE: WebKit currently uses the first shipping option even if other shipping options are marked as selected.
If the requestShipping member is true, then shippingAddress's PaymentAddress must match the expected values.

PASS Can construct a payment request (smoke test).
PASS If the requestShipping member is true, then shippingAddress's PaymentAddress must match the expected values.
